sample main.tf
Here is an example of a main.tf file for a 3-tier architecture:
In this example, the architecture is made up of three tiers: a network tier, an application server tier, and a database server tier. The tiers are defined using Terraform modules, which are stored in the modules directory. The provider block declares the AWS provider, and the var blocks define variables that can be passed to the modules. The resource blocks declare additional AWS resources such as an Amazon RDS instance and an Amazon Elastic Load Balancer (ELB), and associate them with the appropriate components of the network and application server modules.

This example shows how Terraform can be used to manage a 3-tier architecture by breaking down the infrastructure into reusable components and using modules to manage and orchestrate these components. The variables used in this example can be defined in a variables.tf file, and the values for these variables can be stored in a terraform.tfvars file or passed as command line options when you run Terraform.