sample variables.tf
Here is an example of a variables.tf file for the 3-tier architecture:

This file defines the variables that can be passed to the Terraform configuration in the main.tf file. The variables are organized by their purpose, with a description of each variable and its type. The default block sets a default value for each variable in case a value is not provided when Terraform is run. This allows you to parameterize your Terraform configurations and make them more flexible and reusable.