sample outputs.tf
Here is an example of an outputs.tf file for the 3-tier architecture:

This file defines the outputs of the Terraform configuration, which can be useful for debugging and sharing information about the infrastructure. Each output block specifies a name and a value, which is typically a reference to an attribute of an AWS resource.

For example, the app_dns_name output provides the DNS name of the Amazon ELB, which can be used to access the application. The db_endpoint output provides the endpoint for the Amazon RDS instance, which can be used to access the database. The vpc_id output provides the ID of the VPC, and the public_subnets and private_subnets outputs provide the IDs of the public and private subnets, respectively.

By using outputs in Terraform, you can easily share information about your infrastructure with other members of your team or with external systems. You can also use outputs to retrieve information about your infrastructure, such as the DNS name of an ELB, when you run Terraform in a script.