sample terraform.tfvars
Here is an example of a terraform.tfvars file for the 3-tier architecture:

This file is used to store the values for the variables defined in the variables.tf file. Each line in the file corresponds to a variable and its value, separated by an = sign. The values in this file are used to replace the default values in the variables.tf file when Terraform is run.

By using a terraform.tfvars file, you can easily specify the values for your Terraform variables without having to pass them on the command line each time you run Terraform. This makes it easier to manage your Terraform configurations and reduces the risk of errors when specifying values. Additionally, you can use different terraform.tfvars files for different environments, such as development, test, and production, to make it easier to manage your infrastructure in different contexts.